Use float instead of double

+++ start/creep.h
@@ -2,11 +2,11 @@
 #define CREEP_H
 
 /* Walking speed. */
-extern double creep_dx; // Sideways.
-extern double creep_dy; // Forward.
-extern double creep_spread;
-extern double creep_rotation;
-extern double creep_height;
+extern float creep_dx; // Sideways.
+extern float creep_dy; // Forward.
+extern float creep_spread;
+extern float creep_rotation;
+extern float creep_height;
 
 /* Perform one step of the creep gait with the current speed. */
 void walk();
