+++ start/servos.h
@@ -4,7 +4,7 @@
 // Number of servos.
 #define SERVOS 12
 
-void servo_move(unsigned char servo, double rads);
+void servo_move(unsigned char servo, float rads);
 void servo_setup(void);
 void servo_shutdown(void);
